+++ lib/govuk_index.rb
@@ -1,0 +1,9 @@
+module GovukIndex
+  class MissingTextHtmlContentType < StandardError; end
+
+  class NotFoundError < StandardError; end
+
+  class UnknownDocumentTypeError < StandardError; end
+
+  class NotIdentifiable < StandardError; end
+end
