+++ lib/govuk_index/presenters/elasticsearch_presenter.rb
@@ -1,5 +1,7 @@
 module GovukIndex
   class ElasticsearchPresenter
+    class MissingExternalUrl < StandardError; end
+
     include ElasticsearchIdentity
 
     def initialize(payload:, type_mapper:)
